refactor(metatools): move the tool discovery handlers into handlers_tools.go
Pure move, no behaviour change: list_tools, describe_tool, list_core_tools,
filter_tools and the shared filter/rank/paginate engine leave the 1100-line
handlers.go for a file of their own. Prepares the list_tools paging change.
